improvement: remove some mods, add baseq3a mod
The mods that are removed are:
- autoattack, from thelongestyard.link
- ztm-flexible-hud

baseq3a includes the "unlagged" mod,
which should make gameplay better with higher ping.
It still sucks a lot if the connection is not P2P,
i.e. constantly lagging, but at least you don't need to lead
hitscan weapon shots.

Anoter important point of this change is to support other mods.
The changes to the game engine, namely autoattack,
made it incompatible with other mods,
i.e. all player and item models would not load if you try to play
with baseq3a mod added. See
https://github.com/WofWca/quake3.xdc/issues/2#issuecomment-3527641547.