fix: don't use baseq3a mod for demo, use ztm-hud
Its license is not compatible with the demo version.

Partially revert "improvement: remove some mods, add baseq3a mod",
commit 37c777359e0cc80485c7e3b8883d3553bf71c07c.
This commit doesn't fully revert that one.

Unlike prior to that commit, we now use the version
of the ztm-flexible-hud which is based on its original code,
and not the QVM files built from the thelongestyard.link project.
Again, this is because that forked version did some changes
to the engine which are not compatible with any other mods.