fix: don't `defaultBasegame = demoq3` if no demoq3
Currently this bug doesn't really manifest itself,
but it would if we were to make `main-menu.html` the main page.
If we were, then the most prominent manifestation
of the issue that this fixes
would be the issue where after first launch you pick "Play demo"
and then close the app or go back somehow.
Then on next launch of the app you would get
to the "upload demo" screen instead of the main screen.